requirements.txt
attrs==19.3.0
boto3==1.13.15
botocore==1.16.15
cached-property==1.5.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
click-default-group==1.2.2
datapackage==1.14.0
docutils==0.15.2
et-xmlfile==1.0.1
GitPython==3.1.7
goodtables==2.4.16
idna==2.9
ijson==3.0.4
importlib-metadata==1.6.0
isodate==0.6.0
jdcal==1.4.1
jmespath==0.10.0
jsonlines==1.2.0
jsonpointer==2.0
jsonschema==3.2.0
linear-tsv==1.1.0
openpyxl==3.0.3
pyrsistent==0.16.0
python-dateutil==2.8.1
requests==2.23.0
rfc3986==1.4.0
s3transfer==0.3.3
simpleeval==0.9.10
six==1.15.0
SQLAlchemy==1.3.17
statistics==1.0.3.5
tableschema==1.18.0
tabulator==1.48.0
unicodecsv==0.14.1
urllib3==1.25.9
xlrd==1.2.0
zipp==3.1.0
